+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Change: Note compatibility through WP 5.5+
 * Unit tests:
     * New: Add test for hook registrations
+    * New: Add tests for `modify_nextprevious_post_where()`
 
 ## 3.0.1 _(2020-05-10)_
 * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests